Rendering av filer-liste
Mangler:
- Å returnere index eller lignende på onRetry og onDelete sånn at man kan slette én av flere duplikate filer?
- Å laste ned fila når man trykker på den, eventuelt kalle onClick eller rendre a med href
