+++ @navikt/core/react/src/form/file-upload/item/item-context.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
-import { FileItem } from "./Item";
+import { FileItem } from "./props";
 
 export type ItemContextProps = {
   file: FileItem;
