+++ @navikt/core/react/src/form/file-upload/item/props.ts
@@ -1,0 +1,15 @@
+
+export interface FileWithoutInteraction {
+  name: string;
+  size?: number;
+}
+export interface FileWithClick extends FileWithoutInteraction {
+  onClick: () => void
+}
+export interface FileWithLink extends FileWithoutInteraction {
+  href: string;
+}
+
+export type FileMetadata = FileWithLink | FileWithClick | FileWithoutInteraction
+
+export type FileItem = File | FileMetadata
